+++ priv/repo/migrations/20190821214420_add_section_order_to_respondent.exs
@@ -1,0 +1,9 @@
+defmodule Ask.Repo.Migrations.AddSectionOrderToRespondent do
+  use Ecto.Migration
+
+  def change do
+    alter table(:respondents) do
+      add :section_order, :string
+    end
+  end
+end